Refactor Language
I started making some minor rewordings and stuff, and then wanted to switch
over uses of "window size" to "Window_Size" (which is the name that RFC8878
officially give to the parameter), and then it all snowballed into a large
number of edits and a significant refactor of the document.

Let me know what you think, and if you'd like me to pull this apart into
individual changes.